updated pagination and searchForm
