+++ src/pages/Favorites/Favorites.jsx
@@ -4,6 +4,7 @@ import { getFavoriteThunk } from '@/redux/Drinks/operations';
 import { selectFavorites } from '@/redux/Drinks/selectors';
 import Title from '@/shared/components/Title/Title';
 
+import Pagination from '@/shared/components/Pagination/Pagination';
 import { NotAdd } from '@/shared/components/NotAdd/NotAdd';
 import CardList from '@/modules/cardList/components/CardList';
 
@@ -18,6 +19,7 @@ const Favorites = () => {
     <>
       <Title Title="Favorites" />
       {favorites.length > 0 ? <CardList data={favorites} /> : <NotAdd />}
+      <Pagination />
     </>
   );
 };
