+++ src/pages/MyDrinks/MyDrinks.jsx
@@ -5,6 +5,7 @@ import { selectOwn } from '@/redux/Drinks/selectors';
 import { getOwnThunk } from '@/redux/Drinks/operations';
 import { NotAdd } from '@/shared/components/NotAdd/NotAdd';
 import CardList from '@/modules/cardList/components/CardList';
+import Pagination from '@/shared/components/Pagination/Pagination';
 
 const MyDrinks = () => {
   const dispatch = useDispatch();
@@ -18,6 +19,7 @@ const MyDrinks = () => {
     <>
       <Title Title="My drinks" />
       {own.length > 0 ? <CardList data={own} /> : <NotAdd />}
+      <Pagination />
     </>
   );
 };
